test: reorganize tests
