+++ Makefile
@@ -83,7 +83,7 @@ addlicense-install: go-install
 
 pytest: pytest-install
 	cd tests && \
-	$(PYTHON) -m pytest unit --verbose --color=yes --durations=0 \
+	$(PYTHON) -m pytest . --verbose --color=yes --durations=0 \
 		--cov="$(PROJECT_PATH)" --cov-report=xml --cov-report=term-missing
 
 test: pytest
